amplifierd-plugins/amplifierd-plugin-distro/pyproject.toml
[project]
name = "amplifierd-plugin-distro"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["uvicorn[standard]>=0.30"]
test = ["pytest", "httpx"]

[project.entry-points."amplifierd.plugins"]
distro = "distro_plugin"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/distro_plugin"]

[tool.pyright]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"
extraPaths = ["src"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "uvicorn[standard]>=0.30",
]